Order Copy constructor fixed
Before, the copy constructor was giving SIGSEGV error. It was because the Player attribute in it was never initialized

Fix: change Deploy, Advance... default constructor (... p=nullptr) -> (... new Player())

GameEngine, I fixed some minor error
1. checkForWinner():
territoryowner is a string, thus so change to getOwner()->getName()

2. gamePlayTransitions()
issueOrdersPhase() is declare in .h to have no parameter, so when called it must not take in p. remove it